ロジック切り分け 細かいレイアウトの修正
